Only allow images to exist on a single page (#141)
At the moment we list the data for every image with every page, even if they're not used on that page. Make it so that images can exist on only one page. This avoids massive redundancy where information about every image is stored with every page, even when they're not used. Documents with either lots of images, or lots of pages (or both) will benefit most from this.

In theory in the future we might want to allow the same image to exist multiple times, but the API doesn't really allow for that at the moment anyway, so this is simpler.